feat: RND-112: Improve GroundingDINO
