+++ label_studio_ml/examples/grounding_dino/docker-compose.yml
@@ -20,27 +20,21 @@ services:
       - LOG_LEVEL=DEBUG
 
       # Add these variables if you want to access the images stored in Label Studio
-      - LABEL_STUDIO_HOST=
-      - LABEL_STUDIO_ACCESS_TOKEN=
-
-      # use these if you want to use segment anything instead of bounding box predictions from input text prompts
-      - USE_SAM=false  # if you want to automatically generate segment anything model predictions
-      - USE_MOBILE_SAM=false # whether you want to use a more efficient, yet a bit less accurate, version of the segment anything model
+      - LABEL_STUDIO_HOST=https://app.heartex.com/
+      - LABEL_STUDIO_ACCESS_TOKEN=d58b2054f9853fb9152c34f25356a7c90b9efd00
 
       - BOX_THRESHOLD=0.30
       - TEXT_THRESHOLD=0.25
-# # Uncomment the following lines if you want to use GPU
-#      - NVIDIA_VISIBLE_DEVICES=all
-#    deploy:
-#      resources:
-#        reservations:
-#          devices:
-#            - driver: nvidia
-#              count: 1
-#              capabilities: [gpu]
+  # Uncomment the following lines if you want to use GPU
+      - NVIDIA_VISIBLE_DEVICES=all
+    deploy:
+      resources:
+        reservations:
+          devices:
+            - driver: nvidia
+              count: 1
+              capabilities: [gpu]
     ports:
-      - "9090:9090"
+      - "30000:9090"
     volumes:
-      - "./data/ml-backend:/data"
-      - "./prompt.txt:/app/prompt.txt"
-      - ./dino.py:/app/dino.py
+      - "./data/server:/data"
