+++ label_studio_ml/examples/grounding_sam/requirements-test.txt
@@ -1,0 +1,2 @@
+pytest
+pytest-cov
