+++ label_studio_ml/examples/grounding_sam/start.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# Execute the gunicorn command
+exec gunicorn --bind :${PORT:-9090} --workers ${WORKERS:-1} --threads ${THREADS:-4} --timeout 0 _wsgi:app
