+++ label_studio_ml/examples/grounding_sam/test_api.py
@@ -1,0 +1,5 @@
+import pytest
+
+
+def test_example():
+    assert True
